refactor: rewrite resets for v4
