+++ docs/getting-started.md
@@ -2,10 +2,22 @@
 
 ## Prerequisites
 
+Wicara requires `styled-components` v4+. First, install the core Wicara theme/reset package and `styled-components`.
+
+```bash
+# yarn
+$ yarn add styled-components @kata-kit/theme @kata-kit/reset
+
+# npm
+$ npm install --save styled-components @kata-kit/theme @kata-kit/reset
+```
+
 At the bare minimum, you should at least import the global CSS reset provided by Wicara.
 
-```js
-import '@kata-kit/reset';
+```jsx
+import { KataReset } from '@kata-kit/reset';
+
+const App = ({ children }) => <KataReset>{children}</KataReset>;
 ```
 
 ## Example
