+++ packages/reset/src/components/GlobalStyles.ts
@@ -1,448 +1,12 @@
-import { createGlobalStyle, css } from 'styled-components';
-import { variables } from '@kata-kit/theme';
-
-const reboot = css`
-  /*!
-   * Bootstrap Reboot v4.1.2 (https://getbootstrap.com/)
-   * Copyright 2011-2018 The Bootstrap Authors
-   * Copyright 2011-2018 Twitter, Inc.
-   * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)
-   * Forked from Normalize.css, licensed MIT (https://github.com/necolas/normalize.css/blob/master/LICENSE.md)
-   */
-  *,
-  *::before,
-  *::after {
-    box-sizing: border-box;
-  }
-
-  html {
-    font-family: sans-serif;
-    line-height: 1.15;
-    -webkit-text-size-adjust: 100%;
-    -ms-text-size-adjust: 100%;
-    -ms-overflow-style: scrollbar;
-    -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-  }
-
-  ${css`
-    @-ms-viewport {
-      width: device-width;
-    }
-  `} article,
-aside,
-figcaption,
-figure,
-footer,
-header,
-hgroup,
-main,
-nav,
-section {
-    display: block;
-  }
-
-  body {
-    margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-      'Helvetica Neue', Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-      'Segoe UI Symbol';
-    font-size: 1rem;
-    font-weight: 400;
-    line-height: 1.5;
-    color: #212529;
-    text-align: left;
-    background-color: #fff;
-  }
-
-  [tabindex='-1']:focus {
-    outline: 0 !important;
-  }
-
-  hr {
-    box-sizing: content-box;
-    height: 0;
-    overflow: visible;
-  }
-
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    margin-top: 0;
-    margin-bottom: 0.5rem;
-  }
-
-  p {
-    margin-top: 0;
-    margin-bottom: 1rem;
-  }
-
-  abbr[title],
-  abbr[data-original-title] {
-    text-decoration: underline;
-    -webkit-text-decoration: underline dotted;
-    text-decoration: underline dotted;
-    cursor: help;
-    border-bottom: 0;
-  }
-
-  address {
-    margin-bottom: 1rem;
-    font-style: normal;
-    line-height: inherit;
-  }
-
-  ol,
-  ul,
-  dl {
-    margin-top: 0;
-    margin-bottom: 1rem;
-  }
-
-  ol ol,
-  ul ul,
-  ol ul,
-  ul ol {
-    margin-bottom: 0;
-  }
-
-  dt {
-    font-weight: 700;
-  }
-
-  dd {
-    margin-bottom: 0.5rem;
-    margin-left: 0;
-  }
-
-  blockquote {
-    margin: 0 0 1rem;
-  }
-
-  dfn {
-    font-style: italic;
-  }
-
-  b,
-  strong {
-    font-weight: bolder;
-  }
-
-  small {
-    font-size: 80%;
-  }
-
-  sub,
-  sup {
-    position: relative;
-    font-size: 75%;
-    line-height: 0;
-    vertical-align: baseline;
-  }
-
-  sub {
-    bottom: -0.25em;
-  }
-
-  sup {
-    top: -0.5em;
-  }
-
-  a {
-    color: #007bff;
-    text-decoration: none;
-    background-color: transparent;
-    -webkit-text-decoration-skip: objects;
-  }
-
-  a:hover {
-    color: #0056b3;
-    text-decoration: underline;
-  }
-
-  a:not([href]):not([tabindex]) {
-    color: inherit;
-    text-decoration: none;
-  }
-
-  a:not([href]):not([tabindex]):hover,
-  a:not([href]):not([tabindex]):focus {
-    color: inherit;
-    text-decoration: none;
-  }
-
-  a:not([href]):not([tabindex]):focus {
-    outline: 0;
-  }
-
-  pre,
-  code,
-  kbd,
-  samp {
-    font-family: SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono',
-      'Courier New', monospace;
-    font-size: 1em;
-  }
-
-  pre {
-    margin-top: 0;
-    margin-bottom: 1rem;
-    overflow: auto;
-    -ms-overflow-style: scrollbar;
-  }
-
-  figure {
-    margin: 0 0 1rem;
-  }
-
-  img {
-    vertical-align: middle;
-    border-style: none;
-  }
-
-  svg:not(:root) {
-    overflow: hidden;
-    vertical-align: middle;
-  }
-
-  table {
-    border-collapse: collapse;
-  }
-
-  caption {
-    padding-top: 0.75rem;
-    padding-bottom: 0.75rem;
-    color: #6c757d;
-    text-align: left;
-    caption-side: bottom;
-  }
-
-  th {
-    text-align: inherit;
-  }
-
-  label {
-    display: inline-block;
-    margin-bottom: 0.5rem;
-  }
-
-  button {
-    border-radius: 0;
-  }
-
-  button:focus {
-    outline: 1px dotted;
-    outline: 5px auto -webkit-focus-ring-color;
-  }
-
-  input,
-  button,
-  select,
-  optgroup,
-  textarea {
-    margin: 0;
-    font-family: inherit;
-    font-size: inherit;
-    line-height: inherit;
-  }
-
-  button,
-  input {
-    overflow: visible;
-  }
-
-  button,
-  select {
-    text-transform: none;
-  }
-
-  button,
-  html [type='button'],
-  [type='reset'],
-  [type='submit'] {
-    -webkit-appearance: button;
-  }
-
-  button::-moz-focus-inner,
-  [type='button']::-moz-focus-inner,
-  [type='reset']::-moz-focus-inner,
-  [type='submit']::-moz-focus-inner {
-    padding: 0;
-    border-style: none;
-  }
-
-  input[type='radio'],
-  input[type='checkbox'] {
-    box-sizing: border-box;
-    padding: 0;
-  }
-
-  input[type='date'],
-  input[type='time'],
-  input[type='datetime-local'],
-  input[type='month'] {
-    -webkit-appearance: listbox;
-  }
-
-  textarea {
-    overflow: auto;
-    resize: vertical;
-  }
-
-  fieldset {
-    min-width: 0;
-    padding: 0;
-    margin: 0;
-    border: 0;
-  }
-
-  legend {
-    display: block;
-    width: 100%;
-    max-width: 100%;
-    padding: 0;
-    margin-bottom: 0.5rem;
-    font-size: 1.5rem;
-    line-height: inherit;
-    color: inherit;
-    white-space: normal;
-  }
-
-  progress {
-    vertical-align: baseline;
-  }
-
-  [type='number']::-webkit-inner-spin-button,
-  [type='number']::-webkit-outer-spin-button {
-    height: auto;
-  }
-
-  [type='search'] {
-    outline-offset: -2px;
-    -webkit-appearance: none;
-  }
-
-  [type='search']::-webkit-search-cancel-button,
-  [type='search']::-webkit-search-decoration {
-    -webkit-appearance: none;
-  }
-
-  ::-webkit-file-upload-button {
-    font: inherit;
-    -webkit-appearance: button;
-  }
-
-  output {
-    display: inline-block;
-  }
-
-  summary {
-    display: list-item;
-    cursor: pointer;
-  }
-
-  template {
-    display: none;
-  }
-
-  [hidden] {
-    display: none !important;
-  }
-`;
+import { createGlobalStyle } from 'styled-components';
+import reboot from '../styles/reboot';
+import typography from '../styles/typography';
+import base from '../styles/base';
 
 const GlobalStyles = createGlobalStyle`
 ${reboot}
-
-:root {
-  font-size: ${variables.fontProps.rootFontSize};
-  line-height: ${variables.fontProps.rootLineHeight};
-}
-
-html,
-body,
-#root {
-  width: 100%;
-  height: 100%;
-}
-
-body {
-  font-family: 'Museo Sans Rounded', -apple-system, BlinkMacSystemFont,
-    'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji',
-    'Segoe UI Emoji', 'Segoe UI Symbol';
-  background-color: ${variables.colors.gray10};
-  color: ${variables.colors.gray80};
-  font-weight: 500;
-}
-
-a {
-  color: #006fe6;
-  text-decoration: none;
-
-  &:hover {
-    color: #0056b3;
-    text-decoration: underline;
-  }
-
-  &:active {
-    color: #0056b3;
-  }
-
-  &:focus {
-    outline: 2px solid #0056b3;
-    outline-offset: 2px;
-  }
-}
-
-#root {
-  transition: all 0.5s cubic-bezier(0.15, 1, 0.3, 1);
-  -webkit-transition: all 0.5s cubic-bezier(0.15, 1, 0.3, 1);
-
-  &.pushed-legend-right {
-    transform: translateX(-280px);
-  }
-}
-
-.noscroll {
-  overflow: hidden;
-}
-
-.noselect {
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-
-.full-size {
-  height: 100%;
-  width: 100%;
-}
-
-.icon-middle {
-  &::before {
-    vertical-align: middle;
-  }
-}
-
-.drag-handle {
-  cursor: move;
-  display: inline-block;
-
-  &::before {
-    content: '......';
-    display: inline-block;
-    width: 10px;
-    word-break: break-word;
-    white-space: normal;
-    letter-spacing: 2px;
-    line-height: 4.5px;
-    text-align: center;
-    height: 18px;
-  }
-}
+${base}
+${typography}
 `;
 
 export default GlobalStyles;
